move build and cfg packages into util
