+++ clients/cmd/docker-driver/main.go
@@ -14,7 +14,7 @@ import (
 	"github.com/prometheus/common/version"
 	"github.com/weaveworks/common/logging"
 
-	_ "github.com/grafana/loki/pkg/build"
+	_ "github.com/grafana/loki/pkg/util/build"
 )
 
 const socketAddress = "/run/docker/plugins/loki.sock"
