+++ clients/cmd/promtail/main.go
@@ -22,9 +22,9 @@ import (
 	"github.com/grafana/loki/clients/pkg/promtail"
 	"github.com/grafana/loki/clients/pkg/promtail/config"
 
-	_ "github.com/grafana/loki/pkg/build"
-	"github.com/grafana/loki/pkg/cfg"
 	logutil "github.com/grafana/loki/pkg/util"
+	_ "github.com/grafana/loki/pkg/util/build"
+	"github.com/grafana/loki/pkg/util/cfg"
 )
 
 func init() {
