+++ cmd/loki-canary/main.go
@@ -14,10 +14,10 @@ import (
 	"github.com/prometheus/client_golang/prometheus/promhttp"
 	"github.com/prometheus/common/version"
 
-	_ "github.com/grafana/loki/pkg/build"
 	"github.com/grafana/loki/pkg/canary/comparator"
 	"github.com/grafana/loki/pkg/canary/reader"
 	"github.com/grafana/loki/pkg/canary/writer"
+	_ "github.com/grafana/loki/pkg/util/build"
 )
 
 type canary struct {
